Update budget routes
